requirements.txt
asyncio
bleak
numpy
paho-mqtt
